+++ packages/tswebextension/src/lib/mv3/content-script/hideDocumentReferrer.ts
@@ -1,0 +1,3 @@
+import { StealthHelper } from '../../common/stealth-helper';
+
+StealthHelper.hideDocumentReferrer();